requirements.txt
discord.py
python-dotenv
yt_dlp
PyNaCl
aiohttp
mutagen
pytube
setuptools
wheel
